+++ mlxtend/__init__.py
@@ -4,4 +4,4 @@
 #
 # License: BSD 3 clause
 
-__version__ = '0.5.1'
+__version__ = '0.6.1.dev0'