+++ docs/sources/CHANGELOG.md
@@ -6,11 +6,29 @@ The CHANGELOG for the current development version is available at
 [https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md](https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md).
 
 
+### Version 0.6.1 (TBD)
 
-### Version 0.6.0 (TBD)
 
-The CHANGELOG for the current development version is available at
-[https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md](https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md).
+##### Downloads
+
+- [Source code (zip)](https://github.com/rasbt/mlxtend/archive/v0.6.1.zip)
+- [Source code (tar.gz)](https://github.com/rasbt/mlxtend/archive/v0.6.1.tar.gz)
+
+##### New Features
+
+- -
+
+##### Changes
+
+- -
+
+##### Bug Fixes
+
+- -
+
+
+### Version 0.6.0 (2017-03-18)
+
 
 ##### Downloads
 
@@ -19,15 +37,17 @@ The CHANGELOG for the current development version is available at
 
 ##### New Features
 
-- -
+- An `association_rules` function is implemented that allows to generate rules based on a list of frequent itemsets (via [Joshua Goerner](https://github.com/JoshuaGoerner)).
 
 ##### Changes
 
 - Adds a black `edgecolor` to plots via `plotting.plot_decision_regions` to make markers more distinguishable from the background in `matplotlib>=2.0`.
+- The `association` submodule was renamed to `frequent_patterns`.
 
 ##### Bug Fixes
 
-- -
+- The `DataFrame` index of `apriori` results are now unique and ordered.
+- Fixed typos in autompg and wine datasets (via [James Bourbeau](https://github.com/jrbourbeau)).
 
 
 ### Version 0.5.1 (2017-02-14)